Finally the big moment!!! 🌈 Refactor createWebpackConfig into webpack.config.js Now all the webpack is in the main config file
